Add three small changes.
First, I reverted the sqlcmdtest changes for querytimeout.  Second, I
added a test to TestGroupByComplexSuite which causes an assertion
failure in the EE.  It's based on the sqlcmd querytimeout tests.  Third,
I added some code in tabletuple.h to print a backtrace when the
assertion failure is about to happen.